Fix bug
git reset HEAD~5
