--- DJ_Keyboard/DJ_Keyboard.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29306.81
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DJ_Keyboard", "DJ_Keyboard\DJ_Keyboard.vcxproj", "{0543EB69-19A0-43FB-827B-1BF25ACF96B5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Debug|x64.ActiveCfg = Debug|x64
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Debug|x64.Build.0 = Debug|x64
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Debug|x86.ActiveCfg = Debug|Win32
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Debug|x86.Build.0 = Debug|Win32
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Release|x64.ActiveCfg = Release|x64
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Release|x64.Build.0 = Release|x64
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Release|x86.ActiveCfg = Release|Win32
-		{0543EB69-19A0-43FB-827B-1BF25ACF96B5}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D992CFA3-4083-41E9-A250-2BD4A91152EC}
-	EndGlobalSection
-EndGlobal
