+++ keyboard/g_words.txt
@@ -1,0 +1,4 @@
+기러기
+올챙이
+호랑이
+다람쥐
